use `styfle/cancel-workflow-action` for workflow concurrency handling
The built-in concurrency option from GitHub is similar, but it won't allow a new workflow to start until the out-of-date version has finished cancelling.  Job with `if: always()` will execute even after a cancel, which means that the updated branch's workflow will be delayed.  With this action, the new version executes immediately (unless queued...) and cancellation is just done optimistically.